clean up webpack, swap `indexRoute`
This restore HMR. It also swaps the `indexRoute` of the application.
The About page is now the primary view. PostList is now at `/posts`.
