+++ common/routes/About/index.js
@@ -1,6 +1,3 @@
 import About from './components/About'
 
-export default {
-  path: 'about',
-  component: About
-}
+export default About
