+++ common/routes/PostList/index.js
@@ -1,32 +1,19 @@
-import { provideHooks } from 'redial'
-import React, { PropTypes } from 'react'
-import { loadPosts } from './actions'
-import { connect } from 'react-redux'
-import PostListItem from './components/PostListItem'
-import { StyleSheet, css } from 'aphrodite'
+if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
+import { injectAsyncReducer } from '../../store'
 
-const redial = {
-  fetch: ({ dispatch }) => dispatch(loadPosts())
-}
-
-const mapStateToProps = (state) => ({
-  posts: state.posts.data
-})
-
-const PostListPage = ({ posts }) => (
-  <div className={css(styles.root)}>
-    {posts.map((post, i) => <PostListItem key={post.id} post={post} />)}
-  </div>
-)
-
-PostListPage.PropTypes = {
-  posts: PropTypes.array.isRequired
-}
-
-const styles = StyleSheet.create({
-  root: {
-    maxWidth: 500
+export default function createRoutes (store) {
+  return {
+    path: 'posts',
+    getComponents (location, cb) {
+      require.ensure([
+        './containers/PostList',
+        './reducer'
+      ], (require) => {
+        let PostPage = require('./containers/PostList').default
+        let postReducer = require('./reducer').default
+        injectAsyncReducer(store, 'posts', postReducer)
+        cb(null, PostPage)
+      })
+    }
   }
-})
-
-export default provideHooks(redial)(connect(mapStateToProps)(PostListPage))
+}
