+++ common/routes/root.js
@@ -2,7 +2,7 @@
 if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
 
 import App from '../components/App'
-import PostList from './PostList'
+import About from './About'
 
 export default function createRoutes (store) {
   const root = {
@@ -11,7 +11,7 @@ export default function createRoutes (store) {
     getChildRoutes (location, cb) {
       require.ensure([], (require) => {
         cb(null, [
-          require('./About').default, // no need to modify store, no reducer
+          require('./PostList').default(store), // no need to modify store, no reducer
           require('./Post').default(store), // add async reducer
           require('./NotFound').default
         ])
@@ -19,7 +19,7 @@ export default function createRoutes (store) {
     },
 
     indexRoute: {
-      component: PostList
+      component: About
     }
   }
